mach-snapdragon: multi-DTB FIT support
Introduce support for choosing the FDT from a FIT image which is loaded
by the previous stage bootloader as an initramfs.

This works by providing a dummy FDT which the previous stage bootloader
populates with the memory layout, bootargs, and initrd address. That
dummy FDT also contains match data which can be used to pick the correct
FDT from the FIT image.

Initially, matching is done based on the memory layout and/or a
substring in the bootargs. In the future, matching based on SMEM may be
supported as well.

This implementation always takes the first match, so it's important to
order the potential matches from most to least specific.

Signed-off-by: Caleb Connolly <caleb.connolly@linaro.org>